% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winpsw_data.R
\name{winpsw_data}
\alias{winpsw_data}
\title{Generate data with ordinal outcomes and corresponding covariates in a randomized clinical trial}
\usage{
winpsw_data(n, betatrt, betactrl, b01, b02, sds, bernparam, pe)
}
\arguments{
\item{n}{sample size of the data set}

\item{betatrt}{a vector of regression coefficients of linear predictors for ordinal outcome model in treatment group}

\item{betactrl}{a vector of regression coefficients of linear predictors for ordinal outcomes model in control group}

\item{b01}{intercept for ordinal outcome model in treatment group}

\item{b02}{intercept for ordinal outcome model in control group}

\item{sds}{a vector of standard errors for three continuous covariates}

\item{bernparam}{a vector of probabilities for generating three binary covariates}

\item{pe}{the probability for generating treatment variable in a randomized clinical trial}
}
\value{
A data.frame including ordinal outcome, treatment, covariates x1-x6, and quadratic
terms of 3 continuous covariates
}
\description{
This function can generate ordinal outcomes with three levels and 6 covariates with quadratic terms of continuous covariates in a randomized clinical trial
}
\examples{
set.seed(123456)
n = 200
betatrt = c(1,-1,1,-1,1,-1)
betactrl = c(0.5,-0.5,0.5,-0.5,0.5,-0.5)
b01 = 1
b02 = 0.05
sds = c(0.3,0.4,0.5)
bernparam = c(0.75,0.50,0.25)
pe = 0.5
mydata = winpsw_data(n,betatrt,betactrl,b01,b02,sds,bernparam,pe)
print(mydata)

}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}, Scott Zuo \email{scott.zuo@northwestern.edu} and Fan Li \email{fan.f.li@yale.edu}
}
