% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winstat_ipw.R
\name{winstat_ipw}
\alias{winstat_ipw}
\title{Estimate win statistics by IPW method}
\usage{
winstat_ipw(data, outcomevar, treatment, covariate)
}
\arguments{
\item{data}{A dataset of randomized controlled trials (RCT), which should include ordinal outcome, treatment and covariates}

\item{outcomevar}{the name of the outcome in RCT data}

\item{treatment}{the name of the treatment in RCT data}

\item{covariate}{the name of vector covariates used in the propensity score model}
}
\value{
A data.frame containing point and interval estimates of WP, LP, WR and WD
}
\description{
The IPW (i.e., the inverse probability weight) method is a Hajek-type estimator by leveraging U-statistic theory, see section 3 of paper (Cao et al., 2025).
The point estimate is based on formula (9) of paper with IPW weights, and variance estimate is based on formula (15) of paper with IPW weights.
In addition to estimating point estimates of four win statistics, that is, win probability (WP), loss probability (LP),
win ratio (WR) and win difference (WD), their standard errors as well as 95\% confidence intervals will also be returned
}
\examples{
set.seed(123456)
n = 200
betatrt = c(1,-1,1,-1,1,-1)
betactrl = c(0.5,-0.5,0.5,-0.5,0.5,-0.5)
b01 = 1
b02 = 0.05
sds = c(0.3,0.4,0.5)
bernparam = c(0.75,0.50,0.25)
pe = 0.5
mydata = winpsw_data(n,betatrt,betactrl,b01,b02,sds,bernparam,pe)

# example 1 
res1 = winstat_ipw(data=mydata, outcomevar="outcome", treatment="treatment",
       covariate=paste0("x",1:6,seq=""))
print(res1)

}
\references{
Cao Z., Zuo S., Ryan M.M., Davis-Plourde K., Heagerty P., Tong G. and Li F. Covariate-adjusted win statistics in
randomized clinical trials. under review. 2025;0(0):1-20.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}, Scott Zuo \email{scott.zuo@northwestern.edu} and Fan Li \email{fan.f.li@yale.edu}
}
