% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winstat_unadj.R
\name{winstat_unadj}
\alias{winstat_unadj}
\title{Estimate win statistics by unadjusted method}
\usage{
winstat_unadj(data, outcomevar, treatment, method = "IFM")
}
\arguments{
\item{data}{A dataset of randomized controlled trials (RCT), which should include ordinal outcome and treatment.}

\item{outcomevar}{the name of the outcome in RCT data}

\item{treatment}{the name of the treatment in RCT data}

\item{method}{the method used to estimate variance of WP and LP. Two methods are provided, one is the default influence function method (IFM), the other is the Bebu and Lachin method (BLM)}
}
\value{
A data.frame containing point and interval estimates of WP, LP, WR and WD
}
\description{
The unadjusted estimator (i.e., without covariate adjustment) is realized by two methods introduced in paper (Cao et al., 2025). When using the Bebu and Lachin method (BLM),
the point estimate is based on formula (4) of paper, and variance estimate is based on formula (5) of paper. When using the influence function method (IFM),
the point estimate is based on formula (9) of paper with weights =1, and variance estimate is based on formula (18) of paper.
In addition to estimating point estimates of four win statistics, that is, win probability (WP), loss probability (LP),
win ratio (WR) and win difference (WD), their standard errors as well as 95\% confidence intervals will also be returned
}
\examples{
set.seed(123456)
n = 200
betatrt = c(1,-1,1,-1,1,-1)
betactrl = c(0.5,-0.5,0.5,-0.5,0.5,-0.5)
b01 = 1
b02 = 0.05
sds = c(0.3,0.4,0.5)
bernparam = c(0.75,0.50,0.25)
pe = 0.5
mydata = winpsw_data(n,betatrt,betactrl,b01,b02,sds,bernparam,pe)
# example 1: using the default IFM method
res1 = winstat_unadj(data=mydata, outcomevar="outcome", treatment="treatment")
print(res1)
# example 2: Bebu and Lachin method
res2 = winstat_unadj(data=mydata, outcomevar="outcome", treatment="treatment",method="BLM")
print(res2)

}
\references{
Cao Z., Zuo S., Ryan M.M., Davis-Plourde K., Heagerty P., Tong G. and Li F. Covariate-adjusted win statistics in
randomized clinical trials. under review. 2025;0(0):1-20.

Bebu I. and Lachin J. Large sample inference for a win ratio analysis of a composite outcome based on prioritized components.
Biostatistics 2016; 17(1): 178-187.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}, Scott Zuo \email{scott.zuo@northwestern.edu} and Fan Li \email{fan.f.li@yale.edu}
}
